\name{UniversityofTehran}
\alias{UniversityofTehran}
\docType{data}
\title{
Persian texts
}
\description{
Persian text from the University of Tehran website
}
\usage{data("UniversityofTehran")}
\format{
  Persian text data
}
\source{
https://ut.ac.ir/fa/page/200
}
\examples{
# Load data
data(UniversityofTehran)

}
\keyword{datasets}
